Module semester_project::matrix_operations

Modules

matrix_operations_tests 🔒

Functions

back_solve_matrix 🔒

Function to backsolve a matrix in row echelon form

build_augmented_matrix

build an augmented matrix from an x matrix and a y matrix

convert_matrix_to_reduced_row_echelon_form

function to convert a matrix to reduced row echelon form

convert_matrix_to_row_echelon_form

function to convert a matrix to row echelon form

eliminate_row 🔒

Eliminate a row by subtracting each row by a scaled source row

find_largest_row_by_col 🔒

Finds row with largest value finds the index of the row with the largest value at

get_transpose

gets the transpose for a matrix

initialize_result_matrix 🔒

builds a matrix with placeholder values for the rows of left and columns of right

matrix_multiplication

Performs matrix multiplication

scale_row 🔒

Scale a row in a matrix

swap_row 🔒

Swap rows at indexes